move method to composable
